app/requirements.txt
annotated-types==0.7.0
anyio==4.6.0
certifi==2024.8.30
click==8.1.7
dnspython==2.6.1
exceptiongroup==1.2.2
fastapi==0.115.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
motor==3.6.0
packaging==24.1
pluggy==1.5.0
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.9.1
pytest==8.3.3
pytest-asyncio==0.24.0
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.38.6
tomli==2.0.1
typing_extensions==4.12.2
uvicorn==0.30.6
